requirements.txt
# Trip Ethiopia Telegram Bot Dependencies

# Core Telegram Bot
python-telegram-bot==20.7
python-telegram-bot[job-queue]==20.7

# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# API Requests
requests==2.31.0
httpx~=0.25.2
aiohttp==3.9.1

# Currency Conversion
forex-python==1.8

# PDF Generation
reportlab==4.0.9
weasyprint==60.2
pillow==10.2.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2024.1

# Background Tasks
celery==5.3.6
redis==5.0.1

# Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3


